+++ src/tensor_network_viz/einsum_module/__init__.py
@@ -1,11 +1,13 @@
-from ._equation import parse_equation_for_shapes
+from ._equation import parse_einsum_equation, parse_equation_for_shapes
 from .renderer import plot_einsum_network_2d, plot_einsum_network_3d
-from .trace import EinsumTrace, einsum, pair_tensor
+from .trace import EinsumTrace, einsum, einsum_trace_step, pair_tensor
 
 __all__ = [
     "EinsumTrace",
     "einsum",
+    "einsum_trace_step",
     "pair_tensor",
+    "parse_einsum_equation",
     "parse_equation_for_shapes",
     "plot_einsum_network_2d",
     "plot_einsum_network_3d",